feat: split component so it can be used in horizontal feed
